[FEATURE] Add MiddlewareUtil providing convenience functions to all middleware
In order to correctly handle requests with URIs containing escape
sequences like "%20" (=whitespace), middleware implementations need to
decode any escape sequences in incoming request. To make this easier a
new MiddlewareUtil class is added. Every MiddlewareManager creates one
instance of this class and provides it to every standard middleware.

Custom middleware receive an interface to this instance with a set of
methods restricted by the specification version of the custom middleware
extension. This is yet to be implemented in a follow-up change as the
middleware repository does not yet provide specification version
information.