[BREAKING] serveResources middleware: Expect *.properties files in UTF-8 by default
Also fix some tests that expected different unicode escape sequences
for the same input. The used input encoding did not match the specified
propertiesFileSourceEncoding.

BREAKING CHANGE:
If the project a "*.properties" resource originates from cannot be
determined, or if the project does not define a
propertiesFileSourceEncoding configuration or uses a legacy specVersion
(<2.0), the serveResources middleware assumes that the resource is UTF-8
encoded instead of ISO-8859-1.