[FIX] serveResources: Do not process manifest.json in test-resources
The middleware is supposed to be in sync with the behavior of the build.
During the build, only manifest.json files within `/resources/` are
processed, so the server must not process files from `/test-resources/`.

A check for `/resources/` would not be correct, as the pathname is based
on the 'runtime'-style, so for applications the path might just be
`/manifest.json`.